twisted.wamp: add a service for wamp client
This is a common service that twisted applications can use

At the moment, this is very basic, and some more feature would be needed

- automatically reconnect in case some router finds its okay to send random RST on long lasting connections
- not sure about this would work with credentials
